# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nonauth/auth/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User belum aktiv / wajib ganti password"
          }
        },
        "summary": "Login user",
        "tags": [
          "Auth Public"
        ]
      }
    },
    "/nonauth/auth/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOnlyResponseDto"
                }
              }
            }
          }
        },
        "summary": "Refresh token",
        "tags": [
          "Auth Public"
        ]
      }
    },
...
    "/nonauth/auth/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleMessageDto"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth Public"
        ]
      }
    },
...
    "/auth/module/module": {
      "get": {
        "operationId": "ModuleController_findAll",
        "parameters": [
          {
            "name": "pageIndex",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "JSON.stringify({})",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortKey",
            "required": false,
            "in": "query",
            "schema": {
              "example": "created_at",
              "type": "string"
            }
          },
          {
            "name": "sortValue",
            "required": false,
            "in": "query",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search_field",
            "required": false,
            "in": "query",
            "description": "Array of searchable fields",
            "schema": {
              "example": [
                "nama",
                "alamat"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "search_keyword",
            "required": false,
            "in": "query",
            "schema": {
              "example": "makmur",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for list AclModuleDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AclModuleDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "module"
        ]
      },
      "post": {
        "operationId": "ModuleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AclModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create AclModuleDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AclModuleDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "module"
        ]
      }
    },
    "/auth/module/module/{id}": {
      "get": {
        "operationId": "ModuleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get AclModuleDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AclModuleDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "module"
        ]
      },
      "put": {
        "operationId": "ModuleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AclModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put AclModuleDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AclModuleDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "module"
        ]
      },
      "delete": {
        "operationId": "ModuleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete AclModuleDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AclModuleDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "module"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP API",
    "description": "Dokumentasi API ERP multi-tenant",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@email.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AclUsers": {
        "type": "object",
        "properties": {}
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AclUsers"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "TokenOnlyResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "SimpleMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
...
      "AclModuleDto": {
        "type": "object",
        "properties": {
          "idModule": {
            "type": "string"
          },
          "kodeModule": {
            "type": "string"
          },
          "namaModule": {
            "type": "string"
          },
          "namaTabel": {
            "type": "string"
          },
          "namaDb": {
            "type": "string"
          },
          "deskripsi": {
            "type": "string"
          },
          "isAktif": {
            "type": "boolean"
          },
          "urutan": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idModule"
        ]
      },
...
    }
  }
}